requirements.txt
numpy==1.24.2
pandas==1.5.3
Pillow==8.1.2
psutil==5.8.0
pymsteams==0.2.2
pyserial==3.5b0
timeout-decorator==0.5.0